+++ spec/support/servers/checks/passive/insecure_cross_domain_policy_headers.rb
@@ -1,0 +1,13 @@
+require 'sinatra'
+
+get '/crossdomain.xml' do
+    <<EOHTML
+<?xml version="1.0"?>
+<!DOCTYPE cross-domain-policy SYSTEM
+"http://www.adobe.com/xml/dtds/cross-domain-policy.dtd">
+<cross-domain-policy>
+   <site-control permitted-cross-domain-policies="all"/>
+   <allow-http-request-headers-from domain="*" headers="*" secure="false"/>
+</cross-domain-policy>
+EOHTML
+end